backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.7.0

SQLAlchemy==2.0.30
alembic==1.13.1

Werkzeug==3.0.3
Jinja2==3.1.4

bcrypt==4.1.3
python-dotenv==1.0.1

transformers==4.41.2
torch==2.3.0
accelerate==0.31.0
sentencepiece==0.2.0

huggingface-hub==0.23.3
tokenizers==0.19.1
safetensors==0.4.3

langchain==0.2.3
langchain-community==0.2.3

faiss-cpu==1.8.0
sentence-transformers==3.0.0

deep-translator==1.11.4
langdetect==1.0.9

pandas==2.2.2
numpy==1.26.4

tqdm==4.66.4
requests==2.32.3

python-dateutil==2.9.0.post0
pytz==2024.1
reportlab
gunicorn==22.0.0